Build: Error Prone on every Java compile, findings at zero and a finding a build failure (#423, step 3)
error_prone_core 2.50.0 as a javac plugin through maven-compiler-plugin (javac forked so
the jdk.compiler exports it needs stay in this pom); three style checks that rewrite
switch statements and parentheses are off; the javac lint 'restricted' warning is off
(the FFM restricted methods are the point of this code; native access is enabled at
run time). failOnWarning: from here a finding fails the build.

The 60 Error Prone findings and 13 lint warnings of the first run, driven to zero:
- RangeResidual: the operator-mask index no longer assumes compare() returns exactly
  -1/0/1 (CompareToZero) -- an explicit three-way test;
- Bitmap, ColumnBuilder: int arithmetic widened where a long was assigned (IntLongMath);
- StringSliceKernels, BitKernelsTest: the packed low-32-bit lengths masked instead of
  cast to int on the way into a long parameter (UnnecessaryLongToIntConversion);
- unused fields, parameters and a method removed (GroupKeyTable.idxScratch,
  DecimalKernels.n, ArrowCData.I32, WideDecimalKernelsTest.MAX38, ColumnBuilder
  appendUtf8's count, CometVectorAdapter validityOf's numRows, RunKernelsTest check's
  arena, GroupedAccumulators' three-argument regroup);
- MathKernelsTest: a ternary with identical branches; VectorArrowColumnVector: a
  missing @Override; DateKernels: a redundant cast; tests seed Random from the enum
  name, not ordinal();
- Javadoc: duplicate comments merged, a comment on a static block made plain, a
  dangling {@link}, an inline {@code null} that read as the parameter `nulls`;
- suppressed with a reason: three identity comparisons that are caches keyed on the
  object itself (ReferenceEquality), the two enums holding VectorOperators constants
  (ImmutableEnumChecker), the POW10 lookup table (MutablePublicArray), the raw
  Function array in a test.
Kernel tests green.

+++ kernels/src/main/java/io/sparkvector/kernels/Bitmap.java
@@ -55,12 +55,12 @@ public final class Bitmap {
     }
 
     public static void set(MemorySegment bm, int index) {
-        long byteIndex = index >>> 3;
+        int byteIndex = index >>> 3;
         bm.set(BYTE, byteIndex, (byte) (bm.get(BYTE, byteIndex) | (1 << (index & 7))));
     }
 
     public static void clear(MemorySegment bm, int index) {
-        long byteIndex = index >>> 3;
+        int byteIndex = index >>> 3;
         bm.set(BYTE, byteIndex, (byte) (bm.get(BYTE, byteIndex) & ~(1 << (index & 7))));
     }
 
@@ -72,7 +72,6 @@ public final class Bitmap {
         }
     }
 
-    /** Sets or clears the first {@code numBits} bits. */
     /**
      * Sets bits {@code [from, from + count)} to {@code value}; the rest of the
      * bitmap is left as it is.
